backend/requirements.txt
Django==2.2.7
django-cors-headers==3.1.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
Pillow==6.2.1
PyJWT==1.7.1
pytz==2019.3
pycryptodome
